U16 gui windows 2 (#34)
* window logic

* GUI logic

* update

* updated...

* update

* update

* display logic

* update

+++ modules/org.openhs.core.commons/src/org/openhs/core/commons/Door.java
@@ -6,9 +6,9 @@ public class Door extends Thing {
 	public String imagePath_close = "no path";
 	
 	//Coordinates in house 
-	public int x = 0;
-	public int y = 0;
-	public int z = 0;
+	public float x = 0;
+	public float y = 0;
+	public float z = 0;
 	
 	//public boolean open = false;
 	public boolean lock = false;
