+++ modules/org.openhs.core.commons/src/org/openhs/core/commons/TemperatureSensor.java
@@ -1,21 +1,27 @@
+/**
+* @name		TemperatureSensor.java 03/01/2017
+* @author	Michal Valny
+* @version	1.0
+* @description 	Temperature Sensor Data
+* 
+*/
+
 package org.openhs.core.commons;
 
 public class TemperatureSensor extends Thing {
 	
 	//Coordinates in house 
-	public int x = 0;
-	public int y = 0;
-	public int z = 0;
+	public float x = 0.0f;
+	public float y = 0.0f;
+	public float z = 0.0f;
 	
 	Temperature temperature = new Temperature ();
 	
-	public Temperature getTemperature ()
-	{
+	public Temperature getTemperature () {	
 		return temperature;
 	}
 	
-	public void setTemperature (Temperature temp)
-	{
+	public void setTemperature (Temperature temp) {	
 		temperature = temp;
 	}
